pythonApps/xds-config-service/requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
pyyaml==6.0.1
